--- src/main/java/model/Dealer.java
@@ -1,5 +1,0 @@
-package model;
-
-public class Dealer extends Participant{
-    // 카드의 합이 16이하라면 카드를 뽑는다.
-}
