+++ src/main/java/model/card/Suit.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 public enum Suit {
     HEART("하트"),
