+++ src/main/java/model/game/BlackjackGame.java
@@ -1,4 +1,4 @@
-package model;
+package model.game;
 
 public class BlackjackGame {
     // 블랙잭 결과를 비교한다.
