+++ src/main/java/model/participant/Dealer.java
@@ -1,0 +1,5 @@
+package model.participant;
+
+public class Dealer extends Participant {
+    // 카드의 합이 16이하라면 카드를 뽑는다.
+}
