+++ src/main/java/model/participant/Player.java
@@ -1,4 +1,4 @@
-package model;
+package model.participant;
 
 public class Player extends Participant {
     // 카드의 합이 21이하라면 카드를 뽑는다.
