+++ src/main/java/model/participant/Players.java
@@ -1,4 +1,4 @@
-package model;
+package model.participant;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +10,6 @@ public class Players {
     private Players(List<Player> players) {
         this.players = players;
     }
-
 
     public static Players makePlayer(List<String> playerNames) {
         List<Player> playerList = new ArrayList<>();
