+++ src/test/java/model/NameTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.participant.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
